Add support for backing up the journal inode location in the superblock.  E2fsck will automatically save the journal information in the superblock if it is not there already, and will use it if the journal inode appears to be corrupted.  ext2fs_add_journal_inode() will also save the backup information, so that new filesystems created by mke2fs and filesystems that have journals added via tune2fs will also have journal location written to the superblock as well.  Debugfs's logdump command has been enhanced so that it can use the journal information in the superblock.
The debugfs man page has been improved to more fully describe the
logdump command.

Added two new functions, ext2fs_file_open2() and 
ext2fs_inode_io_intern2() which take a pointer to an inode structure;
this is needed so that e2fsck and debugfs can synthesize a
fake journal inode and use it to access the journal.