Debian udeb dependency cleanup
From: Frans Pop <elendil@planet.nl>

Please apply the patch below with your next upload of e2fsprogs.

It removes Provides/Replaces: in the udebs which are no longer needed now
that udeb library dependencies are generated correctly. This has been the
case for a while now, but we've postponed this cleanup until after the
Etch release.

Addresses Debian Bug: #419605

Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>